requirement.txt
aiosqlite==0.17.0
aniso8601==9.0.1
anyio==3.4.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
databases==0.5.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.4
httptools==0.3.0
httpx==0.21.3
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multipledispatch==0.6.0
orjson==3.6.5
outcome==1.1.0
packaging==21.3
passlib==1.7.4
peewee==3.14.8
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
starlette==0.17.1
strawberry-graphql==0.93.23
toml==0.10.2
trio==0.19.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
